Feedback as I follow GoldenPath
Additional Comments:
Line 235 - I received no prompt for naming a binary? I just selected the folder and it built the project files to it.
Line 264 - So at this point, there is nothing for the user to actually see, if I'm understanding this log correctly? Shouldn't we have just a default message in the repo for when each player connects. Then, the user can see something and know that there are 2 players connected in this test. I tried adding a LogInfo to NetworkManager, but it appears the file is immutable and undo's my changes, so I have to extend the NM just to validate what I'm seeing? I got stuck here 
Line 277 - FWIW, this didn't add any log files to my Build directory. It added the file to my current command line location, which I didn't need to 'cd' from since I'm using the "<path to project>"