Update helloworld.md
The testing/what to expect to see when you finished successfully was very hard to distinguish. I didn't really know if it was networked by the end. My open questions were eventually answered by the Golden Path with its extended UI. (Are there 2 player prefabs or just 1? Which window is the host, which is the client?)