Adding more details and renaming to controlled desyncs
I feel like Client Reinterpretation could be confused too easily with Client Prediction in conversations, I looked for a name that was more unique. Controlled Desync still shows the important part of the technique, where we allow clients to be desynced as long as they converge to a common state.